+++ tests/test_multimodule/main2.cpp
@@ -14,8 +14,8 @@ void parse_and_print_2(const char *filename)
     ifstream stream(filename, ios::binary);
     stream.unsetf(ios::skipws);
     stream.seekg(0, ios::end);
-    size_t size = stream.tellg();
-    stream.seekg(0);   
+    size_t size = static_cast<size_t>(stream.tellg());
+    stream.seekg(0);
     vector<char> data(size + 1);
     stream.read(&data.front(), static_cast<streamsize>(size));
     data[size] = '\0';
