+++ tests/test_speed/main.cpp
@@ -133,7 +133,7 @@ void test(const char *filename, const char *description)
         throw runtime_error(string("cannot open file ") + filename);
     stream.unsetf(ios::skipws);
     stream.seekg(0, ios::end);
-    size_t size = stream.tellg();
+    size_t size = static_cast<size_t>(stream.tellg());
     stream.seekg(0);   
     vector<char> data(size + 1);
     stream.read(&data.front(), static_cast<streamsize>(size));
